# swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: 0.1.0
  title: Auction Price Tracker API
host: auction-price-track-server.herokuapp.com
basePath: /api
schemes:
  - https
paths:
...
  /auctions:
    get:
      tags:
        - auctions
      summary: Gets all auctions for a specific item in a realm
      parameters:
        - in: path
          name: realm
          required: true
          type: string
          minimum: 1
          description: the realm to get the auctions from.
        - in: path
          name: itemId
          required: true
          type: integer
          minimum: 2
          description: the item to get the auctions for.
        - in: query
          name: start
          required: false
          type: string
          description: start date parameter for the search. The string must be in the ISO 8601 format. If not specified, it starts 24 hours before the call timestamp
        - in: query
          name: end
          required: false
          type: string
          description: end date parameter for the search. The string must be in the ISO 8601 format. If not specified, it ends at the call timestamp
      operationId: getAuctions
      produces:
        - application/json
      responses:
        '200':
          description: Operation successful
          schema:
            type: array
            items:
              $ref: '#/definitions/AuctionData'
            example:
              - _id: 5c349c376acb5d094ce19c0d
                itemId: 152507
                auctions:
                  - _id: 5c349c376acb5d094ce19c19
                    unitPrice: 337000
                    quantity: 36
                  - _id: 5c349c376acb5d094ce19c18
                    unitPrice: 335000
                    quantity: 5
                  - _id: 5c349c376acb5d094ce19c17
                    unitPrice: 338000
                    quantity: 3
                  - _id: 5c349c376acb5d094ce19c16
                    unitPrice: 385110
                    quantity: 80
                  - _id: 5c349c376acb5d094ce19c15
                    unitPrice: 318182
                    quantity: 55
                  - _id: 5c349c376acb5d094ce19c14
                    unitPrice: 382500
                    quantity: 180
                  - _id: 5c349c376acb5d094ce19c13
                    unitPrice: 350000
                    quantity: 30
                  - _id: 5c349c376acb5d094ce19c12
                    unitPrice: 336000
                    quantity: 13
                  - _id: 5c349c376acb5d094ce19c11
                    unitPrice: 337990
                    quantity: 11
                  - _id: 5c349c376acb5d094ce19c10
                    unitPrice: 998999
                    quantity: 800
                  - _id: 5c349c376acb5d094ce19c0f
                    unitPrice: 339000
                    quantity: 19
                  - _id: 5c349c376acb5d094ce19c0e
                    unitPrice: 399999
                    quantity: 200
                timestamp: '2019-01-08T12:07:32.000Z'
                realm: chogall
                __v: 0
              - _id: 5c349f38e5c26834a4494183
                itemId: 152507
                auctions:
                  - _id: 5c349f38e5c26834a4494191
                    unitPrice: 337000
                    quantity: 36
                  - _id: 5c349f38e5c26834a4494190
                    unitPrice: 335000
                    quantity: 5
                  - _id: 5c349f38e5c26834a449418f
                    unitPrice: 317000
                    quantity: 100
                  - _id: 5c349f38e5c26834a449418e
                    unitPrice: 338000
                    quantity: 3
                  - _id: 5c349f38e5c26834a449418d
                    unitPrice: 385110
                    quantity: 80
                  - _id: 5c349f38e5c26834a449418c
                    unitPrice: 318182
                    quantity: 55
                  - _id: 5c349f38e5c26834a449418b
                    unitPrice: 382500
                    quantity: 180
                  - _id: 5c349f38e5c26834a449418a
                    unitPrice: 316000
                    quantity: 10
                  - _id: 5c349f38e5c26834a4494189
                    unitPrice: 350000
                    quantity: 30
                  - _id: 5c349f38e5c26834a4494188
                    unitPrice: 336000
                    quantity: 13
                  - _id: 5c349f38e5c26834a4494187
                    unitPrice: 337990
                    quantity: 11
                  - _id: 5c349f38e5c26834a4494186
                    unitPrice: 998999
                    quantity: 800
                  - _id: 5c349f38e5c26834a4494185
                    unitPrice: 339000
                    quantity: 19
                  - _id: 5c349f38e5c26834a4494184
                    unitPrice: 399999
                    quantity: 200
                timestamp: '2019-01-08T13:00:32.000Z'
                realm: chogall
                __v: 0
        '500':
          description: Server broken
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unfortunate things happened
  /itemstats:
    get:
      tags:
        - itemstats
      summary: Gets all item statistics for a specific item in a realm
      parameters:
        - in: path
          name: realm
          required: true
          type: string
          minimum: 1
          description: the realm to get the itemstats from.
        - in: path
          name: itemId
          required: true
          type: integer
          minimum: 2
          description: the item to get the itemstats for.
        - in: query
          name: start
          required: false
          type: string
          description: start date parameter for the search. The string must be in the ISO 8601 format. If not specified, it starts 24 hours before the call timestamp
        - in: query
          name: end
          required: false
          type: string
          description: end date parameter for the search. The string must be in the ISO 8601 format. If not specified, it ends at the call timestamp
      operationId: getItemStats
      produces:
        - application/json
      responses:
        '200':
          description: Operation successful
          schema:
            type: array
            items:
              $ref: '#/definitions/ItemStat'
            example:
              - _id: 5c349c396acb5d094ce19d62
                itemId: 152507
                minPrice: 318182
                averagePrice: 723607.3324022347
                maxPrice: 998999
                timestamp: '2019-01-08T12:07:32.000Z'
                realm: chogall
                __v: 0
              - _id: 5c349f39e5c26834a44942dc
                itemId: 152507
                minPrice: 316000
                averagePrice: 694595.1361867704
                maxPrice: 998999
                timestamp: '2019-01-08T13:00:32.000Z'
                realm: chogall
                __v: 0
              - _id: 5c34baf3673b7618dcb5cca7
                itemId: 152507
                minPrice: 336000
                averagePrice: 695249.5205047319
                maxPrice: 998999
                timestamp: '2019-01-08T14:46:32.000Z'
                realm: chogall
                __v: 0
              - _id: 5c34d711673b7618dcb61ff3
                itemId: 152507
                minPrice: 336000
                averagePrice: 695249.5205047319
                maxPrice: 998999
                timestamp: '2019-01-08T16:32:32.000Z'
                realm: chogall
                __v: 0
        '500':
          description: Server broken
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unfortunate things happened
...
definitions:
...
  ItemStat:
    type: object
    properties:
      _id:
        type: string
      itemId:
        type: integer
      minPrice:
        type: integer
      maxPrice:
        type: integer
      averagePrice:
        type: integer
      timestamp:
        type: string
        description: "A timestamp in the ISO Format"
      realm:
        type: string
      __v:
        type: integer
  AuctionData:
    type: object
    properties:
      _id:
        type: string
      itemId:
        type: integer
      auctions:
        type: array
        items:
          type: object
          properties:
            unitPrice:
              type: integer
            quantity:
              type: integer
      timestamp:
        type: string
        description: "A timestamp in the ISO Format"
      realm:
        type: string
      __v:
        type: integer